add: Input Validation and Release Rights Selection
Moved form tag in upload.html up to include the form fields. This is used to enable form validation.

Added a radio input type to select if the selected image's rights belong to the users or not. Based on the selection, the author
and source fields are prefilled if necessary.

Fixes:
* Input validation at frontend
* Release Rights selector

Bug: T258248
Bug: T258247
Bug: T256920
Change-Id: I3dcc65e614728014458ebfedaa2e1d3d7b5e2d28